feat(storage-api): move BlockExecutionWriter trait from provider to storage-api (#15798)
