+++ crates/storage/storage-api/src/lib.rs
@@ -93,3 +93,6 @@ pub use primitives::*;
 
 mod block_indices;
 pub use block_indices::*;
+
+mod block_writer;
+pub use block_writer::*;
